fix(home): Simplify drag end handling and overscroll logic
This commit refactors the drag end handling and overscroll logic across
various screens in the home feature. The `onDragEnd` callbacks are
simplified to not pass a `Float` value, as the velocity information is
no longer directly used for fling animations in this context.

Additionally, the `OffsetOverscrollEffect` has been updated to use a
separate `Animatable` for `overscrollBounceOffset`, allowing for more
nuanced control over bounce effects. The `isInProgress` property now
correctly reflects both overscroll and bounce states. The `node`
implementation within `OffsetOverscrollEffect` has also been refined to
properly apply the `overscrollBounceOffset`.

The affected files include:
- ApplicationScreen.kt
- HorizontalApplicationScreen.kt
- PagerScreenState.kt
- OffsetOverscrollEffect.kt
- OffsetNestedScrollConnection.kt
- VerticalApplicationScreen.kt
- ShortcutConfigScreen.kt
- ListApplicationScreen.kt
- WidgetScreen.kt

+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/screen/shortcutconfig/ShortcutConfigScreen.kt
@@ -135,7 +135,7 @@ internal fun ShortcutConfigScreen(
     onUpdateSharedElementKey: (SharedElementKey?) -> Unit,
     onUpdateIsDragging: (Boolean) -> Unit,
     onVerticalDrag: (Float) -> Unit,
-    onDragEnd: (Float) -> Unit,
+    onDragEnd: () -> Unit,
     onUpdateIsVisibleOverlay: (Boolean) -> Unit,
     onUpdateMoveGridItemResult: (MoveGridItemResult) -> Unit,
 ) {
@@ -309,7 +309,7 @@ private fun EblanShortcutConfigsPage(
     gridItemSettings: GridItemSettings,
     index: Int,
     paddingValues: PaddingValues,
-    onDragEnd: (Float) -> Unit,
+    onDragEnd: () -> Unit,
     onUpdateOverlayBounds: (IntOffset, IntSize) -> Unit,
     onVerticalDrag: (Float) -> Unit,
     onUpdateImageBitmap: (ImageBitmap) -> Unit,
